+++ public/assets/styles/pages/_home.scss
@@ -52,10 +52,6 @@
     text-align: center;
     transition: 0.5s;
 
-    &:hover {
-        cursor: pointer;
-    }
-
     &:hover > .trick__image {
         filter: unset;
     }
