+++ src/Entity/Avatar.php
@@ -18,7 +18,7 @@ class Avatar
 
     #[ORM\OneToOne(inversedBy: 'idAvatar', cascade: ['persist', 'remove'])]
     #[ORM\JoinColumn(nullable: false)]
-    private ?Member $idMember = null;
+    private ?LoginCredentials $idLoginCredentials = null;
 
     public function getId(): ?int
     {
@@ -37,14 +37,14 @@ class Avatar
         return $this;
     }
 
-    public function getIdMember(): ?Member
+    public function getIdLoginCredentials(): ?LoginCredentials
     {
-        return $this->idMember;
+        return $this->idLoginCredentials;
     }
 
-    public function setIdMember(Member $idMember): self
+    public function setIdLoginCredentials(LoginCredentials $idLoginCredentials): self
     {
-        $this->idMember = $idMember;
+        $this->idLoginCredentials = $idLoginCredentials;
 
         return $this;
     }
