+++ templates/base.html.twig
@@ -13,13 +13,13 @@
         <nav class="header__nav">
             <ul class="nav__list">
                 <li class="nav__item">
-                    <a class="link nav__link" href="#">Home</a>
+                    <a class="link nav__link" href="#">Accueil</a>
                 </li>
                 <li class="nav__item">
-                    <a class="link nav__link" href="#">Sign in</a>
+                    <a class="link nav__link" href="#">Se connecter</a>
                 </li>
                 <li class="nav__item">
-                    <a class="link nav__link" href="#">Sign up</a>
+                    <a class="link nav__link" href="#">S'inscrire</a>
                 </li>
             </ul>
         </nav>
@@ -37,6 +37,9 @@
             </a>
             <a class="link link--icon" href="#">
                 <i class="ri-login-box-fill"></i>
+            </a>
+            <a class="link link--icon" href="#">
+                <i class="ri-user-add-line"></i>
             </a>
         </div>
     </footer>
