+++ templates/pages/home.html.twig
@@ -274,7 +274,7 @@
 
         <button class="btn btn__tricks">
             <i class="ri-refresh-line"></i>
-            <span>Load more</span>
+            <span>Charger plus</span>
         </button>
 
         <a class="tricks__arrow link" href="#tricks__list">
